+++ gluecodium/src/test/resources/smoke/external_types/output/cpp/src/smoke/StructWithOverloads.cpp
@@ -9,28 +9,28 @@ namespace smoke {
 static_assert(
     std::is_same<
         int32_t,
-        std::remove_cv<std::remove_reference<decltype(((const external::ClassWithOverloads::StructWithOverloads*)nullptr)->overloadedAccessors())>::type>::type
+        std::remove_cv<std::remove_reference<decltype(std::declval<external::ClassWithOverloads::StructWithOverloads>().overloadedAccessors())>::type>::type
     >::value,
     "Expected 'int32_t' return type for 'external::ClassWithOverloads::StructWithOverloads::overloadedAccessors'."
 );
 static_assert(
     std::is_same<
         void,
-        decltype(((external::ClassWithOverloads::StructWithOverloads*)nullptr)->overloadedAccessors(*((int32_t*)nullptr)))
+        decltype(std::declval<external::ClassWithOverloads::StructWithOverloads>().overloadedAccessors(*((int32_t*)nullptr)))
     >::value,
     "Expected 'void' return type for 'external::ClassWithOverloads::StructWithOverloads::overloadedAccessors'."
 );
 static_assert(
     std::is_same<
         ::std::string,
-        std::remove_cv<std::remove_reference<decltype(((external::ClassWithOverloads::StructWithOverloads*)nullptr)->overloadedMethod())>::type>::type
+        std::remove_cv<std::remove_reference<decltype(std::declval<external::ClassWithOverloads::StructWithOverloads>().overloadedMethod())>::type>::type
     >::value,
     "Expected '::std::string' return type for 'external::ClassWithOverloads::StructWithOverloads::overloadedMethod'."
 );
 static_assert(
     std::is_same<
         ::std::string,
-        std::remove_cv<std::remove_reference<decltype(((external::ClassWithOverloads::StructWithOverloads*)nullptr)->overloadedMethod(
+        std::remove_cv<std::remove_reference<decltype(std::declval<external::ClassWithOverloads::StructWithOverloads>().overloadedMethod(
             *((::std::string*)nullptr)))>::type>::type
     >::value,
     "Expected '::std::string' return type for 'external::ClassWithOverloads::StructWithOverloads::overloadedMethod'."
@@ -38,7 +38,7 @@ static_assert(
 static_assert(
     std::is_same<
         ::std::string,
-        std::remove_cv<std::remove_reference<decltype(((external::ClassWithOverloads::StructWithOverloads*)nullptr)->overloadedMethod(
+        std::remove_cv<std::remove_reference<decltype(std::declval<external::ClassWithOverloads::StructWithOverloads>().overloadedMethod(
             *((::std::string*)nullptr),
             *((bool*)nullptr)))>::type>::type
     >::value,