+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Bug fixes:
+  * Fixed C++ warnings for external types and structs with `@Cpp(Accessors)`.
+
 ## 13.4.1
 Release date: 2022-00-05
 ### Bug fixes: